Move find_matching_endpoint from generate_upload to utils
